+++ qbraid/transpiler/conversions/qasm3/qasm3_to_ionq.py
@@ -18,8 +18,9 @@ from typing import TYPE_CHECKING
 
 from qbraid_core._import import LazyLoader
 
+from qbraid.programs.gate_model.ionq import IONQ_NATIVE_GATES
 from qbraid.transpiler.annotations import weight
-from qbraid.transpiler.conversions.qasm2.qasm2_to_ionq import qasm_to_ionq
+from qbraid.transpiler.conversions.openqasm3.openqasm3_to_ionq import openqasm3_to_ionq
 from qbraid.transpiler.exceptions import CircuitConversionError
 
 pyqasm = LazyLoader("pyqasm", globals(), "pyqasm")
@@ -42,17 +43,19 @@ def qasm3_to_ionq(qasm: Qasm3StringType) -> IonQDictType:
     Raises:
         CircuitConversionError: If the conversion fails.
     """
-    cache_err = None
-
     try:
-        qasm = pyqasm.unroll(qasm)
-    except (ImportError, ModuleNotFoundError) as import_err:
-        cache_err = import_err
-
-    try:
-        return qasm_to_ionq(qasm)
+        return openqasm3_to_ionq(qasm)
     except Exception as err:  # pylint: disable=broad-exception-caught
         err_msg = str(err) or "Failed to convert OpenQASM 3 to IonQ JSON."
+        cache_err = None
+        try:
+            if not any(gate in qasm for gate in IONQ_NATIVE_GATES):
+                qasm = pyqasm.unroll(qasm)
+                return openqasm3_to_ionq(qasm)
+        except NotImplementedError:
+            pass
+        except (ImportError, ModuleNotFoundError) as import_err:
+            cache_err = import_err
         if cache_err and "Cannot mix native and QIS gates in the same circuit" not in err_msg:
             err_msg += " Please install the 'ionq' extra to enable program unrolling with pyqasm."
         else:
