+++ tests/transpiler/qasm/test_qasm_to_ionq.py
@@ -21,11 +21,12 @@ from openqasm3.parser import parse
 
 from qbraid.programs.gate_model.qasm3 import OpenQasm3Program
 from qbraid.programs.typer import IonQDictType, Qasm3StringType
-from qbraid.transpiler.conversions.qasm2.qasm2_to_ionq import (
+from qbraid.transpiler.conversions.openqasm3.openqasm3_to_ionq import (
     _parse_gates,
     extract_params,
-    qasm2_to_ionq,
+    openqasm3_to_ionq,
 )
+from qbraid.transpiler.conversions.qasm2.qasm2_to_ionq import qasm2_to_ionq
 from qbraid.transpiler.conversions.qasm3.qasm3_to_ionq import qasm3_to_ionq
 from qbraid.transpiler.exceptions import CircuitConversionError
 
@@ -347,6 +348,43 @@ def test_qasm3_to_ionq_invalid_params(qasm_code, error_message):
     """Test that qasm3_to_ionq raises an error when the circuit contains invalid parameters."""
     with pytest.raises(CircuitConversionError) as exc_info:
         qasm3_to_ionq(qasm_code)
+    assert error_message in str(exc_info.value)
+
+
+@pytest.mark.parametrize(
+    "qasm_code, error_message",
+    [
+        (
+            """
+    OPENQASM 3.0;
+    qubit[2] q;
+    gpi q[0];
+    """,
+            "Phase parameter is required",
+        ),
+        (
+            """
+    OPENQASM 3.0;
+    qubit[1] q;
+    rz q[0];
+    """,
+            "Angle parameter is required",
+        ),
+        (
+            """
+    OPENQASM 3.0;
+    qubit[2] q;
+    zz q[0], q[1];
+    """,
+            "Angle parameter is required",
+        ),
+    ],
+)
+def test_openqasm3_to_ionq_missing_params(qasm_code, error_message):
+    """Test that openqasm3_to_ionq raises an error when the circuit
+    contains a gate that is missing required parameters."""
+    with pytest.raises(ValueError) as exc_info:
+        openqasm3_to_ionq(qasm_code)
     assert error_message in str(exc_info.value)
